Use size_t for json object array ops

+++ json_object.h
@@ -449,7 +449,7 @@ extern struct array_list* json_object_get_array(struct json_object *obj);
  * @param obj the json_object instance
  * @returns an int
  */
-extern int json_object_array_length(struct json_object *obj);
+extern size_t json_object_array_length(struct json_object *obj);
 
 /** Sorts the elements of jso of type json_type_array
 *
@@ -507,7 +507,7 @@ extern int json_object_array_add(struct json_object *obj,
  * @param idx the index to insert the element at
  * @param val the json_object to be added
  */
-extern int json_object_array_put_idx(struct json_object *obj, int idx,
+extern int json_object_array_put_idx(struct json_object *obj, size_t idx,
 				     struct json_object *val);
 
 /** Get the element at specificed index of the array (a json_object of type json_type_array)
@@ -516,7 +516,7 @@ extern int json_object_array_put_idx(struct json_object *obj, int idx,
  * @returns the json_object at the specified index (or NULL)
  */
 extern struct json_object* json_object_array_get_idx(struct json_object *obj,
-						     int idx);
+						     size_t idx);
 
 /* json_bool type methods */
